Round case 0 option text centering through float
The shipped DrawOptionMenu casts the double centering result to float
(frsp) before truncating to int for the DrawFont2 X coordinate. Wrap
the case 0 expressions in static_cast<float> so the fmadd result is
rounded to single precision (frsp) ahead of the int conversion,
matching the target's frsp/fctiwz sequence.

+++ src/MenuUtil.cpp
@@ -1232,30 +1232,30 @@ void CMenuPcs::DrawOptionMenu()
 			font->SetMargin(kOptionAnimMax);
 			font->SetShadow(1);
 			font->SetScale(firstScale);
-			DrawFont2(static_cast<int>((FLOAT_803335B0 - font->GetWidth(firstText)) * DOUBLE_803335D0 +
-			                           DOUBLE_803335C8),
+			DrawFont2(static_cast<int>(static_cast<float>((FLOAT_803335B0 - font->GetWidth(firstText)) *
+			                           DOUBLE_803335D0 + DOUBLE_803335C8)),
 			          static_cast<int>(FLOAT_803335A4 - FLOAT_803335A0), color, 0x17, firstText, firstScale,
 			          kOptionAnimMax, kOptionAnimMax);
 			font->SetMargin(kOptionAnimMax);
 			font->SetShadow(1);
 			font->SetScale(scale);
-			DrawFont2(static_cast<int>((FLOAT_803335B0 - font->GetWidth(secondText)) * DOUBLE_803335D0 +
-			                           DOUBLE_803335D8),
+			DrawFont2(static_cast<int>(static_cast<float>((FLOAT_803335B0 - font->GetWidth(secondText)) *
+			                           DOUBLE_803335D0 + DOUBLE_803335D8)),
 			          static_cast<int>(FLOAT_803335A4), color, 6, secondText, scale, kOptionAnimMax,
 			          kOptionAnimMax);
 		} else {
 			font->SetMargin(kOptionAnimMax);
 			font->SetShadow(1);
 			font->SetScale(scale);
-			DrawFont2(static_cast<int>((FLOAT_803335B0 - font->GetWidth(firstText)) * DOUBLE_803335D0 +
-			                           DOUBLE_803335C8),
+			DrawFont2(static_cast<int>(static_cast<float>((FLOAT_803335B0 - font->GetWidth(firstText)) *
+			                           DOUBLE_803335D0 + DOUBLE_803335C8)),
 			          static_cast<int>(FLOAT_803335A4), color, 6, firstText, scale, kOptionAnimMax,
 			          kOptionAnimMax);
 			font->SetMargin(kOptionAnimMax);
 			font->SetShadow(1);
 			font->SetScale(scale);
-			DrawFont2(static_cast<int>((FLOAT_803335B0 - font->GetWidth(secondText)) * DOUBLE_803335D0 +
-			                           DOUBLE_803335D8),
+			DrawFont2(static_cast<int>(static_cast<float>((FLOAT_803335B0 - font->GetWidth(secondText)) *
+			                           DOUBLE_803335D0 + DOUBLE_803335D8)),
 			          static_cast<int>(FLOAT_803335A4 - FLOAT_803335A0), color, 0x17, secondText, scale,
 			          kOptionAnimMax, kOptionAnimMax);
 		}
